% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data.R
\name{add_stamps}
\alias{add_stamps}
\title{Add timestamps}
\usage{
add_stamps(df, origin = NULL, unit = "days")
}
\arguments{
\item{df}{a dataframe with datetime column of type date.}

\item{origin}{optional origin (for which stamp is zero)}

\item{unit}{time unit to use for the stamps (delay in ...)}
}
\value{
A dataframe with one more column (named stamp)
}
\description{
Add a timestamps column to df.
}
