% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapHawkes-package.R
\docType{package}
\name{camtrapHawkes-package}
\alias{camtrapHawkes}
\alias{camtrapHawkes-package}
\title{camtrapHawkes: Analyze camera trap data with Hawkes processes}
\description{
Helper functions to infer a Hawkes process from camera trap data. This code relies heavily on the UnitEvents package (https://sourcesup.renater.fr/frs/?group_id=3267) for Hawkes process simulation and inference but implements helper functions for formatting and plotting.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr} (\href{https://orcid.org/0009-0006-5763-0865}{ORCID})

}
\keyword{internal}
