% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{compute_box}
\alias{compute_box}
\title{Compute box}
\usage{
compute_box(d, use_stamps = FALSE)
}
\arguments{
\item{d}{dataframe with columns stamp, datetime and cameraID}

\item{use_stamps}{use timestamps or dates?}
}
\value{
Returns the extimated box (matrix)
  line 1: Camera number
  line 2: start time
  line 3: stop time
  line 4: comportment ID
  each column corresponds to one camera.
}
\description{
Compute time windows covering the whole dataset for a given dataset d.
}
