% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_interevents.R
\name{compute_intervals}
\alias{compute_intervals}
\title{Compute inter-event times}
\usage{
compute_intervals(sp_from, sp_to)
}
\arguments{
\item{sp_from}{species that is observed first}

\item{sp_to}{following species}
}
\value{
Returns the median of the time interval between the 2 
species occurrences
}
\description{
Compute the inter-event times between two species. Function adapted from
Murphy et al 2021 (http://onlinelibrary.wiley.com/doi/abs/10.1111/1365-2656.13548)
}
