% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_interevents.R
\name{compute_intervals_permute}
\alias{compute_intervals_permute}
\title{Compute permuted inter-event times}
\usage{
compute_intervals_permute(sp_from, sp_to)
}
\arguments{
\item{sp_from}{species that is observed first}

\item{sp_to}{following species}
}
\value{
Returns the median of the permuted time interval between the 2 
species occurrences
}
\description{
Compute the permuted inter-event times between species. Function adapted from
Murphy et al 2021 (http://onlinelibrary.wiley.com/doi/abs/10.1111/1365-2656.13548)
}
