% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{compute_pos_neg}
\alias{compute_pos_neg}
\title{Compute true/false positives/negatives}
\usage{
compute_pos_neg(Msimul, Mtrue)
}
\arguments{
\item{Msimul}{a K*K*Ncomp array (K species Ncomp different comportments (often Ncomp =1))}

\item{Mtrue}{true array (K*K*Ncomp array)}
}
\value{
Returns a named vector with four elements named TP, FP, TN, FN
}
\description{
Compute:
    TP (true positives)
    FP (false positives)
    TN (true negatives)
    FN (false negatives)
}
