% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{compute_rate}
\alias{compute_rate}
\title{Compute rates}
\usage{
compute_rate(model, data, timestep = 0.01)
}
\arguments{
\item{model}{a dataframe corresponding to the model estimated with UnitEvents.
This dataframe can be obtained with the function "ue_model_to_df".
Must have columns time, excitefunc, to, from and spont.}

\item{data}{occurrence data. Must have one column stamp and one column species. 
Currently only one camera is supported.}

\item{timestep}{timestep for the function discretisation (days). 
Should be smaller than delta used in the model for good results. Defaults to 0.01.}
}
\value{
A dataframe with columns time, lambda and species.
  time: the time.
  lambda: the intensity function at time t.
  species: the species for which the intensity is computed.
}
\description{
Compute the rate corresponding to data with a given model.
}
