% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{compute_sensi}
\alias{compute_sensi}
\title{Computes sensitivity}
\usage{
compute_sensi(TP, FN)
}
\arguments{
\item{TP}{true positives vector}

\item{FN}{false negatives vector}
}
\value{
The sensitivity (true positive rate)
}
\description{
Computes sensitivity from true positives and false negatives
}
