% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{compute_speci}
\alias{compute_speci}
\title{Computes specificity}
\usage{
compute_speci(TN, FP)
}
\arguments{
\item{TN}{true negatives vector}

\item{FP}{false positives vector}
}
\value{
The specificity (true negative rate)
}
\description{
Computes specificity from true negatives and false positives
}
