% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_interevents.R
\name{compute_TN_TP}
\alias{compute_TN_TP}
\title{Compute TP/TN}
\usage{
compute_TN_TP(pval, true, alpha)
}
\arguments{
\item{pval}{a vector of p-values}

\item{true}{the ground truth vector (must be in the same order as pval)}

\item{alpha}{the significance threshold (defaults to 0.05)}
}
\value{
A named vector TP, FP, TN, FN with the values.
}
\description{
Compute true positive and true negative rates from p-values.
}
