% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{create_funcshape}
\alias{create_funcshape}
\title{Create function shape}
\usage{
create_funcshape(funcshape, times, t = 0.5, strength)
}
\arguments{
\item{funcshape}{exp, gamma or linear}

\item{times}{times vector}

\item{t}{half-life (time value for which the function reaches strength/2).
only used when funcshape == "exp"}

\item{strength}{function max strength coefficient}
}
\value{
A numeric vector of the same length as the times vector containing 
  function values computed for each time.
}
\description{
Create a function shape.
}
