% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{create_interactions}
\alias{create_interactions}
\title{Create interactions}
\usage{
create_interactions(spp_names, times)
}
\arguments{
\item{spp_names}{names of species}

\item{times}{times vector}
}
\value{
A (nspecies, nspecies, 1) array containing in each
  cell a (2, length(times)) matrix: top row is a vector of zero (null)
  interaction function), borrom row is the times vector.
}
\description{
Create an array of null interaction functions
}
