% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{create_interactions_ppstat}
\alias{create_interactions_ppstat}
\title{Create interactions}
\usage{
create_interactions_ppstat(spp_names, times)
}
\arguments{
\item{spp_names}{Species names vector}

\item{times}{Times vector}
}
\value{
A list of lists.
Each list represents one species, and within each list there are the interactions
with this species so that flist[[i]][[j]] represents the effect of j -> i.
}
\description{
Create a list of interactions as needed for a ppstat input filled
with zeroes.
}
\examples{
create_interactions_ppstat(spp_names = letters[1:5], times = seq(0, 10, by = 1))
}
