% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{create_spont_rates}
\alias{create_spont_rates}
\title{Create background rates}
\usage{
create_spont_rates(spont, spp_names)
}
\arguments{
\item{spont}{spontaneous coefficient 
if unique value will be repeated
else will be in the order provided}

\item{spp_names}{names of species}
}
\value{
Returns a matrix of dim (nspecies, 1) containing named spont rates
}
\description{
Create a background rates vector.
}
