% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Transform dataframe to matrix}
\usage{
df_to_matrix(df)
}
\arguments{
\item{df}{A dataframe with columns
cameraID
stamp
species
count}
}
\value{
an array (nspecies, ncameras, nmax+1) in the same format as the output 
  of HawkesMulti from the UnitEvents package). nmax is the maximum number 
  of occurrences for one species on one camera.
}
\description{
Transforms the dataframe into a matrix ready for Lasso inference.
}
