% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data.R
\name{filter_inactive_cameras}
\alias{filter_inactive_cameras}
\title{Filter out cameras}
\usage{
filter_inactive_cameras(df, thr_obs, thr_freq, plot = TRUE)
}
\arguments{
\item{df}{dataframe to filter. Must have columns cameraID, 
snapshotName, count, stamp}

\item{thr_obs}{minimal number of observations to keep camera}

\item{thr_freq}{minimal frequency of observations to keep camera}

\item{plot}{plot graphs?}
}
\value{
Returns the filtered dataframe. If plut = TRUE, ans plots 
camera sampling information.
}
\description{
Filter out cameras that have not enough observations/not enough 
frequent observations
}
