% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{format_data_perf}
\alias{format_data_perf}
\title{Format performance dataframe}
\usage{
format_data_perf(d, add_quantiles = TRUE, level = 0.05)
}
\arguments{
\item{d}{dataframe: must have columns TP, FP, TN, FN.
additional columns possible, but they will affect 
quantiles computation if add_quantiles is TRUE.}

\item{add_quantiles}{should quantiles for sensitivity and specificity be computed?
Quantiles are computed on values grouped by all columns (except "TP", "FP", "TN", "FN").}

\item{level}{level of the quantiles.}
}
\value{
Returns a cleaned form of d:
  additional column sensi computed as TP/(TP+FN)
  additional column speci computed as TN/(TN+FP)
  the columns are pivoted to longer format so that
  sensi and speci are grouped into "value" column and
  "type" describes sensi or speci.
  if add_quantiles, quantile values of value (sensi/speci) will
  be added (quantile computed grouping according to all other columns 
  excluding "TP", "FP", "TN", "FN").
}
\description{
Format a dataframe measuring sensitivity and specificity
}
