% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{get_bins_sum}
\alias{get_bins_sum}
\title{Get bins sum}
\usage{
get_bins_sum(M)
}
\arguments{
\item{M}{an interaction matrix(K*K*Ncomp array)
(K species and Ncomp different comportments (often Ncomp =1))}
}
\value{
A K*K matrix containing 1 if the interaction was inferred, else 0
}
\description{
Get the sum of the bins for a Hawkes model.
}
