% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{get_bins_sum_list}
\alias{get_bins_sum_list}
\title{Get bins sums}
\usage{
get_bins_sum_list(Mlist)
}
\arguments{
\item{Mlist}{Mlist: a list of interaction matrix(K*K*Ncomp array)
(K species and Ncomp different comportments (often Ncomp =1))}
}
\value{
A K*K matrix containing the number of times each value was inferred
}
\description{
Get the sum of the bins for a list of Hawkes model.
}
