% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{get_ids}
\alias{get_ids}
\title{Get IDs}
\usage{
get_ids(df, delete_missing_species = FALSE)
}
\arguments{
\item{df}{a dataframe that must have columns
species
cameraID}

\item{delete_missing_species}{delete species not seen at all cameras?}
}
\value{
a dataframe with columns
   cameraID
   species
   rowid (species_cameraID)
The df is arranged with all species for the same site first
}
\description{
Get a unique row ID for each combination of species/camera.
}
