% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{get_models_df}
\alias{get_models_df}
\title{Get models dataframe}
\usage{
get_models_df(inferred_models, models, est = "BL")
}
\arguments{
\item{inferred_models}{a list with 2 components
   $reinfer: list of models, each one inferred from a different dataset
which is assumed to have been generated by the same true model.
   $reinfer_parameters: metadata about the inferred models}

\item{models}{a list of true models to match the true model that have been 
used to generate data from which models in $reinfer were inferred.}

\item{est}{estimator to use ($BL or $BOL)}
}
\value{
A dataframe  of models
}
\description{
Get a list of all models in inferred_models plus the true model
}
