% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{get_ppstat_coeffs}
\alias{get_ppstat_coeffs}
\title{Get ppstat coefficients}
\usage{
get_ppstat_coeffs(model, term, alpha = 0.05)
}
\arguments{
\item{model}{model object used for ppstat.}

\item{term}{dimension to pick (ie species index)}

\item{alpha}{wanted confidence level}
}
\value{
A dataframe with columns `coef_id`, `species`, `coef`, `lower`
and `upper`.
}
\description{
Returns the coefficients an inferred ppstat model.
}
