% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{get_ppstat_interactions}
\alias{get_ppstat_interactions}
\title{Get ppstat interactions}
\usage{
get_ppstat_interactions(model, alpha = 0.05, trans = NULL)
}
\arguments{
\item{model}{ppstat model (assumed to be multivariate)}

\item{alpha}{confidence level for the confidence intervals (see the documentation of `ppstat::termPlot`)}

\item{trans}{transformation to apply (eg `exp` for a log-intensity) 
(see the documentation of `ppstat::termPlot`)}
}
\value{
A dataframe with columns:

+ `x`: the time
+ `variable`: response variable (species) 
+ `value`: interaction value 
+ `cf.lower`/`cf.upper`: confidence interval bounds
+ `response`: response variable
}
\description{
Get toe interaction functions for ppstat
}
