% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data.R
\name{get_sampling_info}
\alias{get_sampling_info}
\title{Get sampling info}
\usage{
get_sampling_info(df, return = TRUE)
}
\arguments{
\item{df}{A camera trap dataframe. Must have columns
cameraID
stamp}

\item{return}{return the summarized dataframe?}
}
\value{
prints a message summarizing sampling info for the cameras.
}
\description{
Get sampling info
}
