% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{getPlotData}
\alias{getPlotData}
\title{Get plot data}
\usage{
getPlotData(model, ...)
}
\arguments{
\item{model}{ppstat model}

\item{...}{other arguments passed to `my_getTermPlotData` inside the function.}
}
\value{
A dataframe with columns:

+ `x` = time (delay)
+ `variable` = control variables (species from)
+ `value` = value of excitation function
+ `response` = response variable (species to)
}
\description{
Function to return data used for termPlots in ppstat.
Draws heavily on `termPlot` function from `ppstat` package (but
`getTermPlotData` is replaced with `my_getTermPlotData`).
}
