% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{infer}
\alias{infer}
\title{Infer a Hawkes model}
\usage{
infer(
  d,
  k = 12,
  delta = 2/24,
  Z = 0,
  gamma = 0.5,
  scale = 10000,
  use_stamps = FALSE
)
}
\arguments{
\item{d}{dataset to infer from. Must have columns:
snapshotName
cameraID
stamp
datetime
count}

\item{k}{number of bins for the inference function}

\item{delta}{binwidth}

\item{Z}{cutoff}

\item{gamma}{penalization LASSO parameter}

\item{scale}{scale for precision of calculations (defaults to 10000)}

\item{use_stamps}{use timestamps or dates?}
}
\value{
Returns the output in the same form as BoxLasso() function of UnitEvents
A list with 2 (if n(species) = 1) or 3 elements (if n(species) > 1)
  $BL
  $BVL
  $BOL
}
\description{
Infer Hawkes model for data d.
}
