% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{mlist_to_df}
\alias{mlist_to_df}
\title{List of models to dataframes}
\usage{
mlist_to_df(mlist, comp = 1)
}
\arguments{
\item{mlist}{list of models of type UnitEvents 
each component of mlist must be a list with 2 named 
elements $S and $I (corresponds to a model$BL, BOL or $BVL output of BoxLasso)}

\item{comp}{the comportment number to extract}
}
\value{
A dataframe with 4 or 5 columns:
  time: time for the interaction functions
  excitefunc: value of the excitation function
  from: from species
  to: to species
  spont: spontaneous rate of the "to" species
  rep (if mlist is a named list): name of the i-th element of mlist
}
\description{
Transform a list of UnitEvent models to a dataframe
}
