% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{my_getTermPlotData}
\alias{my_getTermPlotData}
\title{get termplot data}
\usage{
my_getTermPlotData(model, alpha = 0.05, trans = NULL)
}
\arguments{
\item{model}{model}

\item{alpha}{confidence level}

\item{trans}{transformation to apply to the results (eg "exp")}
}
\value{
the plot data
}
\description{
Helper function for getPlotData (rewritten from 
getTermPlotData in ppstat package)
}
