% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{name_dataneur}
\alias{name_dataneur}
\title{Name data}
\usage{
name_dataneur(dn, species_names, cameras_names = NA)
}
\arguments{
\item{dn}{an array (nspecies, ncameras, nmax+1) in the same format as the output 
of HawkesMulti from the UnitEvents package).}

\item{species_names}{ordered vector of species names.}

\item{cameras_names}{ordered vector of cameras names.}
}
\value{
The same matrix as the input (dn) but with named rows and columns.
}
\description{
Attribute colnames and rownames to dataneur generated with Hawkesmulti.
}
