% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_background_rate_simu}
\alias{plot_background_rate_simu}
\title{Plot background rates for simulation}
\usage{
plot_background_rate_simu(df, alpha = 0.05, title = NA)
}
\arguments{
\item{df}{the dataframe with true and estimated intensities. It has columns:
rep: repetition ID (s... for simul, true for true data)
excitefunc: value of the excitation function
from/to: ID of the species
time: time in days}

\item{alpha}{confidence level for plotting}

\item{title}{optional plot title}
}
\value{
A ggplot of the background rates where the true rate 
is in red and the inferred rate in blue with a confidence 
interval (if there were several inferred models.)
}
\description{
Plots the inferred background rates vs the real rates
}
