% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot bias}
\usage{
plot_bias(bias_df, fill = "valprop", textsize = 12)
}
\arguments{
\item{bias_df}{a dataframe containing the summarized results of the 
inference for several repetitions. Must have columns:
from and to (interacting species)
the column specified with 'fill'}

\item{fill}{name of the column to use for values to fill the plot.}

\item{textsize}{size of the text}
}
\value{
a ggplot tile with colors corresponding to the proportion
of times the corresponding interaction was inferred.
}
\description{
Plots the bias for each interaction from several inferences.
}
