% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot graph}
\usage{
plot_graph(
  g,
  layout = c(),
  repel = FALSE,
  coledges = "grey",
  colnodes = "cadetblue3",
  coltext = "black",
  textsize = 5,
  s = 8,
  arrsize = 6,
  nudge_x = NA,
  parse_labels = FALSE,
  use_labels_column = FALSE
)
}
\arguments{
\item{g}{The graph to plot (tbl_graph object)}

\item{layout}{optional layout (defaults to layout_in_circle)}

\item{repel}{repel node labels?}

\item{coledges}{color of edges (a string color name)}

\item{colnodes}{color of nodes (named vector named as species or a string
color name). If it is NULL, species will be colored automatically.}

\item{coltext}{color of text for the node labels. It can be a vector
(will correspond to species alphabetical order),}

\item{textsize}{node label size}

\item{s}{size of nodes}

\item{arrsize}{size of arrow}

\item{nudge_x}{optional vector to move x labels for node labels.}

\item{parse_labels}{parse text labels?}

\item{use_labels_column}{use column 'label' for nodes labels?}
}
\value{
A ggplot object representing the graph.
}
\description{
Plots a graph g.
}
