% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot interaction functions}
\usage{
plot_interactions(
  ue_df,
  scale = "days",
  title = NA,
  relative = FALSE,
  silhouettes = NA,
  timestep = NA,
  ystep = NA,
  textsize = 10,
  linesize = 0.5,
  line_geom = c("step", "line"),
  confint = NULL,
  col_conf = "cornflowerblue",
  col_baseline = "black",
  alpha_conf = 0.7,
  h = 0.3,
  baseline = 0,
  separate_self = FALSE
)
}
\arguments{
\item{ue_df}{dataframe with results of UnitEvents inference. Must have columns:

+ time
+ excitefunc
+ from
+ to}

\item{scale}{days or hours, following whether we want the time axis
graduated in days or hours}

\item{title}{plot title}

\item{relative}{plot intensity absolute or relative (divided by spont) value?}

\item{silhouettes}{optional labels with animal silhouettes to replace default labels.
If it exists, must be a named vector with each name 
corresponding to a species name in ue_df. The elements are 
markdown codes containing a <img src='pato_to_image'/> element.}

\item{timestep}{optional timestep for x-axis.}

\item{ystep}{optional step for y-axis (function values).}

\item{textsize}{text minimal size (for x and y axes)}

\item{linesize}{linewidth}

\item{line_geom}{The `geom` to use for the interaction function line: `geom_step`
(adapted for `UnitEvents`) or `geom_line`.}

\item{confint}{If you want confidence intervals around the value of the interaction function, 
use this argument. The first element is the name of the column containing the lowest bound 
of the confidence interval and the second element is the highest bound.
If `NULL` (the default) no confidence interval is plotted.}

\item{col_conf}{Color of the confidence interval}

\item{col_baseline}{Color of the line used to show the baseline}

\item{alpha_conf}{Transparency of the confidence interval}

\item{h}{Horizontal spacing between plots (if separate self)}

\item{baseline}{Baseline against which to compare interaction functions.
Defaults to zero (adapted for UnitEvents model).}

\item{separate_self}{whether to separate auto-interactions and plot them above}
}
\value{
ggplot object, a plot with the pairwise interaction functions between species.
}
\description{
Plot UnitEvents interaction functions
}
