% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_interactions_simu}
\alias{plot_interactions_simu}
\title{Plot interaction function for simulation}
\usage{
plot_interactions_simu(
  df,
  title = NA,
  textsize = 10,
  baseline = 0,
  timestep = NA,
  line_geom = c("step", "line"),
  col_true = "brown3",
  col_infer = "darkblue",
  col_conf = "cornflowerblue",
  conf = c("sample", "confint"),
  col_baseline = "black",
  show.legend = FALSE,
  parse = FALSE,
  y_angle = 90,
  alpha_conf = 0.5,
  level = 0.05
)
}
\arguments{
\item{df}{a dataframe containing values for the inferred function(s) and the true function. 
It has columns:

+ `rep`: repetition ID (s... for simul, true for true data)
+ `excitefunc`: value of the excitation function
+ `from/to`: ID of the species
+ `time`: time in days
+ if `conf = confint`: `cf.lower` and `cf.upper` with the pre-computed confidence intervals.}

\item{title}{optional plot title}

\item{textsize}{text minimal size (for x and y axes)}

\item{baseline}{Baseline against which to compare interaction functions.
Defaults to zero (adapted for UnitEvents model).}

\item{timestep}{optional timestep for x-axis.}

\item{line_geom}{The `geom` to use for the interaction function line: `geom_step`
(adapted for `UnitEvents`) or `geom_line`.}

\item{col_true}{Color of the true function}

\item{col_infer}{Color of the inferred function}

\item{col_conf}{Color of the confidence interval around the inferred function}

\item{conf}{The type of confidence interval to display: either the confidence interval
computed as the quantiles of the inferred functions (if `sample`) or the median of
the pre-computed confidence interval (`confint`). The second option is useful when 
the model was inferred with `ppstat`. If `conf = confint`, 
`df` must have two additional columns `cf.lower` and `cf.upper`.}

\item{col_baseline}{color of the line for the baseline to compare the function to.}

\item{show.legend}{Show the legend?}

\item{parse}{parse the facet labels?}

\item{y_angle}{Angle for the text in y in the panels}

\item{alpha_conf}{Transparency of the confidence interval around the inferred function}

\item{level}{confidence level to use around simulations}
}
\value{
A ggplot of the interaction functions where the true function 
is in red and the inferred function in blue with a confidence 
interval (if there were several inferred models.)
}
\description{
Plots the inferred interaction functions vs the real function
}
