% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_observed_rate}
\alias{plot_observed_rate}
\title{Plot rates}
\usage{
plot_observed_rate(
  rates,
  data,
  timestep = 2,
  textsize = 20,
  ptsize = 1,
  lwd = 1,
  t1,
  t2,
  hlambda,
  hpoints,
  minor_spacing,
  major_spacing,
  max_lambda,
  ybreaks,
  cols,
  ylabel = TRUE,
  xlab = "Time (days)"
)
}
\arguments{
\item{rates}{a dataframe with pre-computed rates. Must have columns:
time
lambda
species}

\item{data}{the corresponding occurrence data. Must have columns:
stamp  
species}

\item{timestep}{timestep to plot}

\item{textsize}{base text size (axes text, 
axes labels are a 2 units bigger)}

\item{ptsize}{point sizes for the events}

\item{lwd}{linewidth for intensities}

\item{t1}{minimal time bounds for subsetting data. If missing, then all data will be plotted.}

\item{t2}{maximal time bounds for subsetting data. If missing, then all data will be plotted.}

\item{hlambda}{plot relative heights for intensity panels}

\item{hpoints}{plot relative heights for points panels}

\item{minor_spacing}{space between the intensity function 
and the correponding points.}

\item{major_spacing}{space between points and the following intensity.}

\item{max_lambda}{max value for y-axis for rates}

\item{ybreaks}{breaks for y-axis (for lambda plots)}

\item{cols}{named vector for colors: names are species names and contain colors}

\item{ylabel}{display labels ?}

\item{xlab}{xlabel to display (optional)}
}
\value{
A ggplot object generated with patchwork.
  Multiple plots in the same column, where all plots are paired,
  the top plot representing the intensity and the bottom plot the actual
  occurrences for one species.
}
\description{
Plot the rate of an observed sequence of events.
}
