% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{reinfer}
\alias{reinfer}
\title{Reinfer Hawkes model}
\usage{
reinfer(M, Ntrial, k = 12, delta = 2/24, gamma = 0.5, scale = 10000)
}
\arguments{
\item{M}{dataset to infer from, a DataNeur matrix (output of HawkesMulti
from the UnitEvents package). It is an array (nspecies, ncameras, nmax+1)
where nmax is the maximum number of occurrences for one species on
one camera.}

\item{Ntrial}{number of cameras}

\item{k}{number of bins for the interaction functions}

\item{delta}{binwidth}

\item{gamma}{penalization LASSO parameter}

\item{scale}{scale for precision of calculations (defaults to 10000)}
}
\value{
Returns the output in the same form as BoxLasso() function of UnitEvents
A list with 2 (if n(species) = 1) or 3 elements (if n(species) > 1)
  $BL
  $BVL
  $BOL
}
\description{
Reinfer Hawkes model from a dataset generated with a Hawkes model.
}
