% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_compendium.R
\name{require}
\alias{require}
\title{Install and load a package (if not already loaded)}
\usage{
require(x)
}
\arguments{
\item{x}{The package name}
}
\value{
Installs the package if it was not already installed
}
\description{
Install and load a package (if not already loaded)
}
