% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data.R
\name{shift_duplicates}
\alias{shift_duplicates}
\title{Shift duplicates}
\usage{
shift_duplicates(d, precision = 10^6)
}
\arguments{
\item{d}{a dataframe with columns
row_ID (row id)
cameraID (camera identifier)
stamp (timestamp, numeric)}

\item{precision}{precision to consider}
}
\value{
input dataframe with modified timestamps.
}
\description{
Shift records occuring at the same timestamp on the same camera
}
