% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{ue_model_to_df}
\alias{ue_model_to_df}
\title{Model to dataframe}
\usage{
ue_model_to_df(mod, comp = 1)
}
\arguments{
\item{mod}{a list of length 2 (and is the named output of the UnitEvent's
function coeff2interac):
$S (matrix (n(species), Ncomp)).
$I (array (n(species), n(species), Ncomp)).}

\item{comp}{the comportment to extract}
}
\value{
a dataframe with columns
  time: the time (support of interaction functions)
  excitefunc: values of the functions for a corresponding time t
  from: species "from" which the interaction is
  to: species "towards" which the interaction is
}
\description{
Transforms the specification of a UnitEvents model 
to a dataframe.
}
