% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{ue_to_graph}
\alias{ue_to_graph}
\title{Model to graph}
\usage{
ue_to_graph(IS, neurnames)
}
\arguments{
\item{IS}{a list of length 2 (the output of the UnitEvent's function coeff2interac):
$S (matrix (n(species), Ncomp)).
$I (array (n(species), n(species), Ncomp)).}

\item{neurnames}{species_names}
}
\value{
a tbl_graph object representing the model.
  nodes: 1 column: names (with species names).
  edge: 3 columns: from, to, weight.
}
\description{
Adaptation of plot_graph_Hawkes function in UnitEvents, 
that returns the igraph object instead of plotting it.
}
