% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_unitevents.R
\name{unpack_inf}
\alias{unpack_inf}
\title{Format the inference output}
\usage{
unpack_inf(inf, species_names = NA, k, delta)
}
\arguments{
\item{inf}{the output of the infer or BoxLasso. Must be a list with 3 elements:
$BL
$BVL
$BOL}

\item{species_names}{a list of species names to rename rows and columns (optional)}

\item{k}{number of bins}

\item{delta}{binwidth}
}
\value{
A list of length 3:
  $BL
  $BVL
  $BOL
Each element is a list of length 2 (and is the named output of the UnitEvent's
  function coeff2interac):
  $S (matrix (n(species), Ncomp)). The matrix's rows are named
    with species_names.
  $I (array (n(species), n(species), Ncomp)). The array's first
  and second dimensions are named with species_names.
}
\description{
Transform output of inference into a list of coefficient matrices.
}
