% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppstat.R
\name{write_formula}
\alias{write_formula}
\title{Write formula}
\usage{
write_formula(
  spp,
  startknot_spp = -0.5,
  endknot_spp = 2.5,
  by_spp = 0.25,
  ord = 3,
  hourcov = FALSE,
  ord_hour = 3,
  startknot_hour = -pi/4,
  endknot_hour = 2 * pi + pi/4,
  by_hour = (3/24) * 2 * pi,
  trunc = NULL
)
}
\arguments{
\item{spp}{Vector of unique species names}

\item{startknot_spp}{First knot for species interactions. It is expressed
in the same unit as the stamps on the data to fit the model to (e.g. days).}

\item{endknot_spp}{Last knot for species interactions. It is expressed
in the same unit as the stamps on the data to fit the model to (e.g. days).}

\item{by_spp}{Step for the species interactions knots. It is expressed
in the same unit as the stamps on the data to fit the model to (e.g. days).}

\item{ord}{Order of the splines of species interactions.}

\item{hourcov}{Should the final formula include a hour covariate (TRUE or FALSE)?}

\item{ord_hour}{Order of the splines of the hour covariate for background rates.}

\item{startknot_hour}{First knot for the hour covariate spline for the background rate
(expressed in radians).}

\item{endknot_hour}{Last knot for the hour covariate spline for the background rate
(expressed in radians).}

\item{by_hour}{Step of the knots for the hour covariate spline for the background rate
(expressed in radians).}

\item{trunc}{Whether to truncate the species interactions. If it is NULL, species interactions splines will
not be truncated. Else, it must be a numerical vectors and the spline basis will be truncated to 
the interval from trunc[1] to trunc[2] (see the documentation of ppstat::bSpline).}
}
\value{
A string that can be interpreted as a formula.
}
\description{
Write a formula expected by ppstat for the inference model
}
