% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_background_rate}
\alias{plot_background_rate}
\title{Plot background rates}
\usage{
plot_background_rate(
  ue_df,
  title = NA,
  textsize = 10,
  silhouettes = NA,
  write_label = FALSE,
  nudge_label = 0.3
)
}
\arguments{
\item{ue_df}{ue_df: dataframe with results of UnitEvents inference. Must have columns:
time
excitefunc
from
to}

\item{title}{plot title}

\item{textsize}{text minimal size (for x and y axes)}

\item{silhouettes}{optional labels with animal silhouettes to replace default labels.
If it exists, must be a named vector with each name 
corresponding to a species name in ue_df. The elements are 
markdown codes containing a <img src='pato_to_image'/> element.}

\item{write_label}{write background rates values besides the points?}

\item{nudge_label}{if the background rates are written, by how much should they 
be nudged on the x-axis?}
}
\value{
a ggplot object, representing background rates for each species.
}
\description{
Plot the background rate for species.
}
