% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_interactions_simu}
\alias{plot_interactions_simu}
\title{Plot interaction function for simulation}
\usage{
plot_interactions_simu(df, title = NA, level = 0.05)
}
\arguments{
\item{df}{a dataframe containing values for the inferred function(s)
and the true function. It has columns:
rep: repetition ID (s... for simul, true for true data)
excitefunc: value of the excitation function
from/to: ID of the species
time: time in days}

\item{title}{optional plot title}

\item{level}{confidence level to use around simulations}
}
\value{
A ggplot of the interaction functions where the true function 
is in red and the inferred function in blue with a confidence 
interval (if there were several inferred models.)
}
\description{
Plots the inferred interaction functions vs the real function
}
